requirements.txt
llama-index==0.14.10
llama-index-readers-file==0.5.6
llama-index-vector-stores-chroma==0.5.3
llama-index-embeddings-openai==0.5.1
llama-index-llms-anthropic==0.10.4
llama-index-llms-openai==0.6.12
langchain-experimental==0.4.1
langchain-openai==1.1.6
fastapi==0.127.0
uvicorn==0.40.0
chromadb==1.3.7
pypdf==6.5.0
python-docx==1.2.0
cohere==5.20.1
openai==2.14.0
anthropic==0.75.0
python-multipart==0.0.21
rank-bm25==0.2.2
python-dotenv==1.2.1
docx2txt==0.9
docx2python
olefile